Shuffle the pattern match code around.
