+++ opencog/query/VirtualMatch.cc
@@ -246,94 +246,6 @@ bool PatternMatch::recursive_virtual(PatternMatchCallback *cb,
 }
 
 /* ================================================================= */
-
-/**
- * Validate a collection of clauses and negations for correctness.
- *
- * Every clause should contain at least one variable in it; clauses
- * that are constants and can be trivially discarded.
- * Furthermore, all clauses should be connected. Two clauses are
- * connected if they contain a common variable.
- *
- * As a side effect, this looks for 'virtual links' and separates
- * them out into a distinct list, _virtuals and _nonvirts.
- *
- * It also partition the clauses into a set of connected components,
- * _components.
- */
-void PatternMatch::validate_clauses(std::set<Handle>& vars,
-                                    std::vector<Handle>& clauses)
-
-{
-	// Make sure that the user did not pass in bogus clauses.
-	// Make sure that every clause contains at least one variable.
-	// The presence of constant clauses will mess up the current
-	// pattern matcher.  Constant clauses are "trivial" to match,
-	// and so its pointless to even send them through the system.
-	bool bogus = remove_constants(vars, clauses);
-	if (bogus)
-	{
-		logger().warn("%s: Constant clauses removed from pattern matching",
-		              __FUNCTION__);
-	}
-
-	// Make sure that each declared variable appears in some clause.
-	// We can't ground variables that aren't attached to something.
-	// Quoted variables are constants, and so don't count.
-	for (Handle v : vars)
-	{
-		if (not is_variable_in_any_tree(clauses, v))
-		{
-			std::stringstream ss;
-			ss << "The variable " << v->toString()
-			   << " does not appear (unquoted) in any clause!";
-			throw InvalidParamException(TRACE_INFO, ss.str().c_str());
-		}
-	}
-
-	// Are there any virtual links in the clauses? If so, then we need
-	// to do some special handling.  BTW: a clause is virtual only if
-	// it contains a GroundedPredicate, and the GroundedPredicate takes
-	// an argument that contains a variable. Otherwise, its not really
-	// virtual.
-	//
-	// The GreaterThanLink is a link type that implicitly contains
-	// a GroundedPredicate for numeric greater-than relations. So
-	// we search for that too.
-	//
-	// XXX FIXME, the check below is not quite correct; for example,
-	// it would tag the following as virtual, although it is not:
-	// (BlahLink (VariableNode "$var") (EvaluationLink (GPN "scm:duh")
-	// (ListLink (ConceptNode "Stuff"))))  -- the var is there but not
-	// in the GPN.
-	for (Handle clause: clauses)
-	{
-		if ((contains_atomtype(clause, GROUNDED_PREDICATE_NODE)
-		    or contains_atomtype(clause, GREATER_THAN_LINK))
-		    and any_variable_in_tree(clause, vars))
-			_virtuals.push_back(clause);
-		else
-			_nonvirts.push_back(clause);
-	}
-
-#ifdef I_DONT_THINK_THIS_CHECK_IS_NEEDED
-	// For now, the virtual links must be at the top. Not sure
-	// what code breaks if this isn't the case.  Why are we checking
-	// this???
-	for (Handle v : _virtuals)
-	{
-		Type vt = v->getType();
-		if ((not classserver().isA(vt, EVALUATION_LINK))
-		    and (not classserver().isA(vt, GREATER_THAN_LINK)))
-		{
-			throw InvalidParamException(TRACE_INFO,
-				"Expecting EvaluationLink at the top level!");
-		}
-	}
-#endif
-}
-
-/* ================================================================= */
 /**
  * Ground (solve) a pattern; perform unification. That is, find one
  * or more groundings for the variables occuring in a collection of
